+++ src/main/java/com/github/smurd/command/CommandContainer.java
@@ -16,17 +16,23 @@ public class CommandContainer {
     private final ImmutableMap<String, Command> commandMap;
     private final Command unknownCommand;
 
-    public CommandContainer(SendBotMessageService sendBotMessageService, TelegramUserService telegramUserService,
-                            DrummerSubService drummerSubService, DrummerRepository drummerRepository) {
+    public CommandContainer(SendBotMessageService sendBotMessageService,
+            TelegramUserService telegramUserService, DrummerSubService drummerSubService,
+            DrummerRepository drummerRepository) {
 
         commandMap = ImmutableMap.<String, Command>builder()
-                .put(START.getCommandName(), new StartCommand(sendBotMessageService, telegramUserService))
-                .put(STOP.getCommandName(), new StopCommand(sendBotMessageService, telegramUserService))
+                .put(START.getCommandName(), new StartCommand(sendBotMessageService,
+                        telegramUserService))
+                .put(STOP.getCommandName(), new StopCommand(sendBotMessageService,
+                        telegramUserService))
                 .put(HELP.getCommandName(), new HelpCommand(sendBotMessageService))
                 .put(NO.getCommandName(), new NoCommand(sendBotMessageService))
-                .put(STAT.getCommandName(), new StatCommand(sendBotMessageService, telegramUserService))
-                .put(ADD_DRUMMER_SUB.getCommandName(), new AddDrummerSubCommand(sendBotMessageService,
-                        drummerSubService, drummerRepository))
+                .put(STAT.getCommandName(), new StatCommand(sendBotMessageService,
+                        telegramUserService))
+                .put(ADD_DRUMMER_SUB.getCommandName(), new AddDrummerSubCommand(
+                        sendBotMessageService, drummerSubService, drummerRepository))
+                .put(LIST_DRUMMER_SUB.getCommandName(), new ListDrummerSubCommand(
+                        sendBotMessageService, telegramUserService))
                 .build();
 
         unknownCommand = new UnknownCommand(sendBotMessageService);